PythonProject1/requirements.txt
aiogram==3.20.0.post0
python-decouple == 3.8
aiohttp == 3.11.18